add beer_order , beer_order_line
@ManyToOne @OneToMany

+++ spring6andTesting/src/main/resources/db/migration/V1__init-mysql-database.sql
@@ -6,7 +6,7 @@
     create table beer (
        id varchar(36) not null,
         beer_name varchar(50) not null,
-        beer_style smallint not null,
+        beer_style TINYINT not null,
         created_date datetime(6),
         price decimal(38,2) not null,
         quantity_on_hand integer,
